+++ DAL/ViewModels/CustomerViewModel.cs
@@ -6,7 +6,7 @@ public class CustomerViewModel
 
     public string CustomerName { get; set; } = null!;
 
-    public int? Phoneno { get; set; }
+    public long? Phoneno { get; set; }
 
     public string? Email { get; set; }
