+++ DAL/ViewModels/OrderDetaIlsInvoiceViewModel.cs
@@ -24,7 +24,7 @@ public class OrderDetaIlsInvoiceViewModel
 
     public string CustomerName { get; set; } = null!;
 
-    public int? Phoneno { get; set; }
+    public long? Phoneno { get; set; }
 
     public long NumberOfPerson{get;set;}
